requirements.txt
flask==3.0.0
python-dotenv==1.0.0
twilio==7.16.0
google-generativeai==0.8.1
gunicorn==21.2.0